+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/SDKConfiguration.java
@@ -36,8 +36,8 @@ class SDKConfiguration {
 	public String language = "java";
 	public String openapiDocVersion = "0.1.0";
 	public String sdkVersion = "2.1.1";
-	public String genVersion = "2.188.3";
-	public String userAgent = "speakeasy-sdk/java 2.1.1 2.188.3 0.1.0 <no value>";
+	public String genVersion = "2.225.2";
+	public String userAgent = "speakeasy-sdk/java 2.1.1 2.225.2 0.1.0 <no value>";
 	public Map<String, Map<String, Map<String, Object>>> globals = new HashMap<String, Map<String, Map<String, Object>>>(){{
 		put("parameters", new HashMap<String, Map<String, Object>>());
 	}};