+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/models/operations/IgnoredGenerationGetResponse.java
@@ -53,8 +53,9 @@ public class IgnoredGenerationGetResponse {
         return this;
     }
     
-    public IgnoredGenerationGetResponse(@JsonProperty("ContentType") String contentType, @JsonProperty("StatusCode") Integer statusCode) {
+    public IgnoredGenerationGetResponse(@JsonProperty("ContentType") String contentType, @JsonProperty("StatusCode") Integer statusCode, @JsonProperty("RawResponse") HttpResponse<byte[]> rawResponse) {
         this.contentType = contentType;
         this.statusCode = statusCode;
+        this.rawResponse = rawResponse;
   }
 }