+++ python-client-sdk/src/sdk/sdkconfiguration.py
@@ -41,8 +41,8 @@ class SDKConfiguration:
     language: str = 'python'
     openapi_doc_version: str = '0.1.0'
     sdk_version: str = '3.1.2'
-    gen_version: str = '2.188.3'
-    user_agent: str = 'speakeasy-sdk/python 3.1.2 2.188.3 0.1.0 openapi'
+    gen_version: str = '2.225.2'
+    user_agent: str = 'speakeasy-sdk/python 3.1.2 2.225.2 0.1.0 openapi'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: