+++ typescript-client-sdk/src/sdk/models/operations/multiplesimpleschemeauth.ts
@@ -31,5 +31,5 @@ export class MultipleSimpleSchemeAuthResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }