+++ typescript-client-sdk/src/sdk/models/operations/typedparametergenerationget.ts
@@ -48,5 +48,5 @@ export class TypedParameterGenerationGetResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }