+++ RELEASES.md
@@ -1052,3 +1052,21 @@ Based on:
 - [typescript v2.1.2] typescript-client-sdk
 ### Releases
 - [Go v2.1.2] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v2.1.2 - go-client-sdk
+
+## 2024-01-05 00:12:59
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 
+- Speakeasy CLI 1.130.1 (2.225.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v2.1.2] go-client-sdk
+- [java v2.1.1] java-client-sdk
+- [php v2.1.1] php-client-sdk
+- [python v3.1.2] python-client-sdk
+- [ruby v2.1.1] ruby-client-sdk
+- [swift v0.3.1] swift-client-sdk
+- [typescript v2.1.2] typescript-client-sdk
+- [csharp v0.3.1] csharp-client-sdk
+### Releases
+- [Go v2.1.2] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v2.1.2 - go-client-sdk
+- [Swift Package Manager v0.3.1] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/swift-client-sdk/v0.3.1 - swift-client-sdk